Refactor: useUserInfo를 useAuthService와 합치기
